backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pillow
google-generativeai
python-multipart
# optional: any other packages you explicitly need
motor
